service-map微服务：上传车辆终端、搜索车辆终端以及解析String转换成Integer，Integer强转成Long的精度损失问题

+++ internal-common/src/main/java/com/mashibing/internalcommon/response/TerminalResponse.java
@@ -14,5 +14,5 @@ public class TerminalResponse {
 	 */
 	private Integer tid;
 
-	private Integer desc;
+	private Long desc;
 }
